Move the shared reflector primitives to reflector.rs
Verdict, ReplyRewrite, and NoRewrite were split across simple.rs and
search.rs but are used by both (and the protocol modules). Move all three
to the shared parent reflector.rs, next to the other cross-reflector
infrastructure (egress_sources, BuildError). simple.rs and search.rs each
now hold only their reflector, depending one-directionally on the parent.
